+++ src/entrypoints/clubcentral-done/main.ts
@@ -1,0 +1,3 @@
+import { startCountdown } from "../../shared/countdown";
+
+startCountdown();